% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2021_01_01}
\alias{min_bic_2021_01_01}
\title{a nationwide collection of best state-specific AR/MA model orders, 12/28}
\format{
S3 instance
}
\usage{
min_bic_2021_01_01
}
\description{
a nationwide collection of best state-specific AR/MA model orders, 12/28
}
\keyword{datasets}
