% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2021_12_05}
\alias{min_bic_2021_12_05}
\title{a nationwide collection of best state-specific AR/MA model orders, 12/05/21}
\format{
S3 instance
}
\usage{
min_bic_2021_12_05
}
\description{
a nationwide collection of best state-specific AR/MA model orders, 12/05/21
}
\keyword{datasets}
