% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncapp.R
\name{plot_sync}
\alias{plot_sync}
\title{display synchronized death and confirmed case series}
\usage{
plot_sync(
  dfr,
  basedate = "2020-05-01",
  maxdate = Inf,
  shift = 26,
  winsize = 14,
  na_pad = TRUE
)
}
\arguments{
\item{dfr}{data.frame with columns 'date', 'deaths', 'confirmed', where the latter are 'incident',
formed by differencing the cumulative counts provided in sars2pack}

\item{basedate}{minimum date to consider}

\item{maxdate}{maximum date to use, Inf for no limit relative to available dates}

\item{shift}{number of days to backshift death counts, coerced to integer}

\item{winsize}{size of running window for averaging incident counts for plotting}

\item{na_pad}{logical(1) passed to runner::mean_run}
}
\description{
display synchronized death and confirmed case series
}
\note{
The 'IFR' is crudely estimated as exp(mean(log(deaths)-log(confirmed))).
}
\examples{
if (!exists("jh")) jh = enriched_jhu_data()
requireNamespace("dplyr")
us = dplyr::filter(jh, alpha2Code=="US")
usd =  dplyr::filter(us, subset=="deaths")
usc =  dplyr::filter(us, subset=="confirmed")
usd2 = usd[-1,]
usd2$inc = diff(usd$count)
usc2 = usc[-1,]
usc2$inc = diff(usc$count)
uss = rbind(usd2, usc2)
uscd = data.frame(date=usc2$date, confirmed=usc2$inc, deaths=usd2$inc)
plot_sync(uscd)
}
