% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob.R
\name{plot_sync_mob}
\alias{plot_sync_mob}
\title{illustrate joint behavior of mobility data and COVID-19 incidence over time in states}
\usage{
plot_sync_mob(state = "California", shift = 0)
}
\arguments{
\item{state}{character(1) capitalized}

\item{shift}{numeric(1) optional, would add a trace if > 0}
}
\description{
illustrate joint behavior of mobility data and COVID-19 incidence over time in states
}
\examples{
op = par(no.readonly=TRUE)
par(mfrow=c(2,2), mar=c(4,3,3,2))
plot_sync_mob()
plot_sync_mob("New York")
plot_sync_mob("Massachusetts")
plot_sync_mob("Texas")
par(op)
}
