% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncapp.R
\name{resync}
\alias{resync}
\title{shift the death records in dfr \code{shift} days into the past}
\usage{
resync(dfr, shift = 26)
}
\arguments{
\item{dfr}{a data.frame with date, deaths, confirmed}

\item{shift}{numeric(1) coerced to integer}
}
\description{
shift the death records in dfr \code{shift} days into the past
}
\examples{
ind = data.frame(date=1:5, confirmed=letters[1:5], deaths=LETTERS[1:5])
resync(ind, 2)
}
