% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncapp.R
\name{syncapp}
\alias{syncapp}
\title{shiny app for appraisal of synchronization of death incidence and confirmed infection incidence peaks}
\usage{
syncapp()
}
\description{
shiny app for appraisal of synchronization of death incidence and confirmed infection incidence peaks
}
